SHA512PaddingStandard

+++ CryptoCommon/IPaddingStandard.cs
@@ -4,7 +4,7 @@ namespace CryptoCommon
 {
     public interface IPaddingStandard
     {
-        IEnumerable<IEnumerable<byte>> GetAlignedBlocks(IEnumerable<byte> lastBlock, int blockSize, int previousBlocksCount);
+        IEnumerable<IEnumerable<byte>> GetAlignedBlocks(IEnumerable<byte> lastBlock, int blockSize, long previousBlocksCount);
         IEnumerable<byte> GetData(IEnumerable<byte> lastBlock, int blockSize);
     }
 }
